clean up code and fix build for macOS and FreeBSD

+++ src/process_group.c
@@ -21,7 +21,6 @@
 
 #include <string.h>
 #include <stdlib.h>
-#include <limits.h>
 #include <sys/time.h>
 #include <signal.h>
 #include <libgen.h>
@@ -32,27 +31,27 @@
 #include "process_group.h"
 #include "list.h"
 
-// look for a process by pid
-// search_pid   : pid of the wanted process
-// return:  pid of the found process, if successful
-//          negative pid, if the process does not exist or if the signal fails
+/* look for a process by pid
+search_pid   : pid of the wanted process
+return:  pid of the found process, if successful
+		 negative pid, if the process does not exist or if the signal fails */
 int find_process_by_pid(pid_t pid)
 {
 	return (kill(pid, 0) == 0) ? pid : -pid;
 }
 
-// look for a process with a given name
-// process: the name of the wanted process. it can be an absolute path name to the executable file
-//         or just the file name
-// return:  pid of the found process, if it is found
-//         0, if it's not found
-//         negative pid, if it is found but it's not possible to control it
+/* look for a process with a given name
+process: the name of the wanted process. it can be an absolute path name to the executable file
+		or just the file name
+return:  pid of the found process, if it is found
+		0, if it's not found
+		negative pid, if it is found but it's not possible to control it */
 int find_process_by_name(const char *process_name)
 {
-	// pid of the target process
+	/* pid of the target process */
 	pid_t pid = -1;
 
-	// process iterator
+	/* process iterator */
 	struct process_iterator it;
 	struct process proc;
 	struct process_filter filter;
@@ -61,15 +60,15 @@ int find_process_by_name(const char *process_name)
 	init_process_iterator(&it, &filter);
 	while (get_next_process(&it, &proc) != -1)
 	{
-		// process found
+		/* process found */
 		if (strcmp(basename(proc.command), process_name) == 0)
 		{
 			if (kill(proc.pid, 0) == -1 && errno == EPERM)
 			{
-				// do not have permission
+				/* do not have permission */
 				return -1;
 			}
-			// process is ok!
+			/* process is ok! */
 			pid = proc.pid;
 			break;
 		}
@@ -78,19 +77,19 @@ int find_process_by_name(const char *process_name)
 		exit(1);
 	if (pid >= 0)
 	{
-		// ok, the process was found
+		/* ok, the process was found */
 		return pid;
 	}
 	else
 	{
-		// process not found
+		/* process not found */
 		return 0;
 	}
 }
 
 int init_process_group(struct process_group *pgroup, int target_pid, int include_children)
 {
-	// hashtable initialization
+	/* hashtable initialization */
 	memset(&pgroup->proctable, 0, sizeof(pgroup->proctable));
 	pgroup->target_pid = target_pid;
 	pgroup->include_children = include_children;
@@ -109,7 +108,7 @@ int close_process_group(struct process_group *pgroup)
 	{
 		if (pgroup->proctable[i] != NULL)
 		{
-			// free() history for each process
+			/* free() history for each process */
 			destroy_list(pgroup->proctable[i]);
 			free(pgroup->proctable[i]);
 			pgroup->proctable[i] = NULL;
@@ -121,24 +120,12 @@ int close_process_group(struct process_group *pgroup)
 	return 0;
 }
 
-void remove_terminated_processes(struct process_group *pgroup)
-{
-	// TODO
-}
+/* returns t1-t2 in millisecond */
+/* static inline double timediff_in_ms(const struct timeval *t1, const struct timeval *t2) */
+#define timediff_in_ms(t1, t2) \
+	(((t1)->tv_sec - (t2)->tv_sec) * 1e3 + ((t1)->tv_usec - (t2)->tv_usec) / 1e3)
 
-// return t1-t2 in microseconds (no overflow checks, so better watch out!)
-// static inline unsigned long timediff(const struct timeval *t1, const struct timeval *t2)
-// {
-// 	return (t1->tv_sec - t2->tv_sec) * 1000000 + (t1->tv_usec - t2->tv_usec);
-// }
-
-// returns t1-t2 in millisecond
-static inline double timediff_in_ms(const struct timeval *t1, const struct timeval *t2)
-{
-	return (t1->tv_sec - t2->tv_sec) * 1e3 + (t1->tv_usec - t2->tv_usec) / 1e3;
-}
-
-// parameter in range 0-1
+/* parameter in range 0-1 */
 #define ALFA 0.08
 #define MIN_DT 20
 
@@ -148,10 +135,10 @@ void update_process_group(struct process_group *pgroup)
 	struct process tmp_process;
 	struct process_filter filter;
 	struct timeval now;
+	double dt;
 	gettimeofday(&now, NULL);
-	// time elapsed from previous sample (in ms)
-	// long dt = timediff(&now, &pgroup->last_update) / 1000;
-	double dt = timediff_in_ms(&now, &pgroup->last_update);
+	/* time elapsed from previous sample (in ms) */
+	dt = timediff_in_ms(&now, &pgroup->last_update);
 	filter.pid = pgroup->target_pid;
 	filter.include_children = pgroup->include_children;
 	init_process_iterator(&it, &filter);
@@ -160,15 +147,12 @@ void update_process_group(struct process_group *pgroup)
 
 	while (get_next_process(&it, &tmp_process) != -1)
 	{
-		//		struct timeval t;
-		//		gettimeofday(&t, NULL);
-		//		printf("T=%ld.%ld PID=%d PPID=%d START=%d CPUTIME=%d\n", t.tv_sec, t.tv_usec, tmp_process.pid, tmp_process.ppid, tmp_process.starttime, tmp_process.cputime);
 		int hashkey = pid_hashfn(tmp_process.pid);
 		if (pgroup->proctable[hashkey] == NULL)
 		{
-			// empty bucket
-			pgroup->proctable[hashkey] = malloc(sizeof(struct list));
+			/* empty bucket */
 			struct process *new_process = malloc(sizeof(struct process));
+			pgroup->proctable[hashkey] = malloc(sizeof(struct list));
 			tmp_process.cpu_usage = -1;
 			memcpy(new_process, &tmp_process, sizeof(struct process));
 			init_list(pgroup->proctable[hashkey], 4);
@@ -177,11 +161,11 @@ void update_process_group(struct process_group *pgroup)
 		}
 		else
 		{
-			// existing bucket
+			/* existing bucket */
 			struct process *p = (struct process *)locate_elem(pgroup->proctable[hashkey], &tmp_process);
 			if (p == NULL)
 			{
-				// process is new. add it
+				/* process is new. add it */
 				struct process *new_process = malloc(sizeof(struct process));
 				tmp_process.cpu_usage = -1;
 				memcpy(new_process, &tmp_process, sizeof(struct process));
@@ -190,21 +174,22 @@ void update_process_group(struct process_group *pgroup)
 			}
 			else
 			{
+				double sample;
 				assert(tmp_process.pid == p->pid);
 				assert(tmp_process.starttime == p->starttime);
 				add_elem(pgroup->proclist, p);
 				if (dt < MIN_DT)
 					continue;
-				// process exists. update CPU usage
-				double sample = 1.0 * (tmp_process.cputime - p->cputime) / dt;
+				/* process exists. update CPU usage */
+				sample = (tmp_process.cputime - p->cputime) / dt;
 				if (p->cpu_usage == -1)
 				{
-					// initialization
+					/* initialization */
 					p->cpu_usage = sample;
 				}
 				else
 				{
-					// usage adjustment
+					/* usage adjustment */
 					p->cpu_usage = (1.0 - ALFA) * p->cpu_usage + ALFA * sample;
 				}
 				p->cputime = tmp_process.cputime;
@@ -220,9 +205,10 @@ void update_process_group(struct process_group *pgroup)
 int remove_process(struct process_group *pgroup, int pid)
 {
 	int hashkey = pid_hashfn(pid);
+	struct list_node *node;
 	if (pgroup->proctable[hashkey] == NULL)
-		return 1; // nothing to delete
-	struct list_node *node = (struct list_node *)locate_node(pgroup->proctable[hashkey], &pid);
+		return 1; /* nothing to delete */
+	node = (struct list_node *)locate_node(pgroup->proctable[hashkey], &pid);
 	if (node == NULL)
 		return 2;
 	delete_node(pgroup->proctable[hashkey], node);
